+++ src/utils/calculateDamage/utils/getEvsText.ts
@@ -1,0 +1,3 @@
+export const getEvsText = (evs: { [key: string]: number }) => {
+  return `${evs.hp}/${evs.atk}/${evs.def}/${evs.spa}/${evs.spd}/${evs.spe}`;
+};
