+++ src/utils/calculateDamage/utils/getPercentDamage.ts
@@ -1,0 +1,4 @@
+export const getPercentDamage = (num: number, maxHP: number) => {
+  const percent = (num / maxHP) * 100 ?? 0;
+  return percent;
+};
